Tidy up methods only used by cli
