+++ src/nexgen/command_line/cli_utils.py
@@ -11,14 +11,11 @@ import h5py
 import numpy as np
 from freephil.common import scope_extract as ScopeExtract  # Define scope extract type
 from numpy.typing import ArrayLike
+from scanspec.core import Path as ScanPath
+from scanspec.specs import Line
 
 from .. import reframe_arrays
-from ..nxs_write import (
-    calculate_scan_range,
-    find_grid_scan_axes,
-    find_number_of_images,
-    find_osc_axis,
-)
+from ..nxs_write import find_number_of_images
 from ..nxs_write.NXclassWriters import (
     write_NXdata,
     write_NXdatetime,
@@ -34,6 +31,213 @@ from ..tools.DataWriter import generate_event_files, generate_image_files
 from ..tools.MetaReader import overwrite_beam, overwrite_detector
 from ..tools.VDS_tools import image_vds_writer, vds_file_writer
 from ..utils import get_filename_template, units_of_time
+
+
+def find_osc_axis(
+    axes_names: List,
+    axes_starts: List,
+    axes_ends: List,
+    axes_types: List,
+    default: str = "omega",
+) -> str:
+    """
+    Identify the rotation scan_axis.
+
+    This function identifies the scan axis from the list passed as argument.
+    The scan axis is the one where start and end value are not the same.
+    If there is only one rotation axis, that is the one returned.
+    In the case scan axis cannot be identified, a default value is arbitrarily assigned.
+
+    Args:
+        axes_names (List): List of names associated to goniometer axes.
+        axes_starts (List): List of start values.
+        axes_ends (List): List of end values.
+        axes_types (List): List of axes types, useful to identify only the rotation axes.
+        default (str, optional): String to deafult to in case scan axis is not found. Defaults to "omega".
+
+    Raises:
+        ValueError: If no axes have been passed.
+        ValueError: If more than one rotation axis seems to move.
+
+    Returns:
+        scan_axis (str): String identifying the rotation scan axis.
+    """
+    # This assumes that at least one rotation axis is always passed.
+    # Assuming all list are of the same length ...
+    if len(axes_names) == 0:
+        raise ValueError(
+            "Impossible to determine translation scan. No axes passed to find_osc_axis function. Please make sure at least one value is passed."
+        )
+    # assert len(axes_names) > 0, "Please pass at least one axis."
+    # Look only for rotation axes
+    rot_idx = [i for i in range(len(axes_types)) if axes_types[i] == "rotation"]
+    axes_names = [axes_names[j] for j in rot_idx]
+    axes_starts = [axes_starts[j] for j in rot_idx]
+    axes_ends = [axes_ends[j] for j in rot_idx]
+
+    if len(axes_names) == 1:
+        scan_axis = axes_names[0]
+    else:
+        idx = [(i != j) for i, j in zip(axes_starts, axes_ends)]
+        if idx.count(True) == 0:
+            # just in case ...
+            scan_axis = default
+        elif idx.count(True) == 1:
+            scan_axis = axes_names[idx.index(True)]
+        else:
+            raise ValueError("Unable to correctly identify the rotation scan axis.")
+    return scan_axis
+
+
+def find_grid_scan_axes(
+    axes_names: List,
+    axes_starts: List,
+    axes_ends: List,
+    axes_types: List,
+) -> List[str]:
+    """
+    Identify the scan axes for a linear/grid scan.
+
+    Args:
+        axes_names (List): List of names associated to goniometer axes.
+        axes_starts (List): List of start values.
+        axes_ends (List): List of end values.
+        axes_types (List): List of axes types, useful to identify only the translation axes.
+
+    Raises:
+        ValueError: If no axes have been passed.
+
+    Returns:
+        scan_axis (List[str]): List of strings identifying the linear/grid scan axes. If no axes are identified, it will return an empty list.
+    """
+    if len(axes_names) == 0:
+        raise ValueError(
+            "Impossible to determine translation scan. No axes passed to find_grid_scan_axes function. Please make sure at least one value is passed."
+        )
+
+    # Look only at translation axes
+    grid_idx = [i for i in range(len(axes_names)) if axes_types[i] == "translation"]
+    axes_names = [axes_names[j] for j in grid_idx]
+    axes_starts = [axes_starts[j] for j in grid_idx]
+    axes_ends = [axes_ends[j] for j in grid_idx]
+
+    scan_axis = []
+    for n, ax in enumerate(axes_names):
+        if axes_starts[n] != axes_ends[n]:
+            scan_axis.append(ax)
+    return scan_axis
+
+
+def calculate_scan_range(
+    axes_names: List,
+    axes_starts: List,
+    axes_ends: List,
+    axes_increments: List = None,
+    n_images: Tuple | int = None,
+    snaked: bool = True,
+    rotation: bool = False,
+) -> Dict[str, ArrayLike]:
+    """
+    Calculate the scan range for a linear/grid scan or a rotation scan from the number of images to be written.
+    If the number of images is not provided, it can be calculated from the increment value of the axis (these values are mutually exclusive).
+    When dealing with a rotation axis, if there are multiple images but no rotation scan, return axis_start repeated n_images times.
+
+    Args:
+        axes_names (List): List of names for the axes involved in the scan.
+        axes_starts (List): List of axis positions at the beginning of the scan.
+        axes_ends (List): List of axis positions at the end of the scan.
+        axes_increments (List, optional): List of ranges through which the axes move each frame. Mostly used for rotation scans. Defaults to None.
+        n_images (Tuple | int, optional): Number of images to be written. If writing a 2D scan, it should be a (nx, ny) tuple, \
+                                        where tot_n_img=nx*ny, any int value is at this time ignored. Defaults to None.
+        snaked (bool): If True, scanspec will "draw" a snaked grid. Defaults to True.
+        rotation (bool): Tell the function to calculate a rotation scan. Defaults to False.
+
+    Raises:
+        TypeError: If the input axes are not lists.
+        ValueError: When an empty axes names list has been passed.
+        ValueError: When both axes_increments and n_images have been passed. The two values are mutually exclusive.
+        ValueError: When neither axes_increments not n_images have been passed.
+        ValueError: For a grid scan, if axes_increments is None, n_images must be a tuple of len=2 to be sure to accurately calculate the scan points.
+
+    Returns:
+        Dict[str, ArrayLike]: A dictionary of ("axis_name": axis_range) key-value pairs.
+    """
+    if type(axes_names) != list or type(axes_starts) != list or type(axes_ends) != list:
+        raise TypeError("Input values for axes must be passed as lists.")
+
+    if len(axes_names) == 0:
+        raise ValueError("No axes have been passed, impossible to determine scan.")
+
+    if n_images and axes_increments:
+        raise ValueError(
+            "The axes_increments and n_images arguments are mutually exclusive. Please pass just one of those."
+            "For a 2D scan it is recommended that n_images is passed."
+        )
+    elif not n_images and not axes_increments:
+        raise ValueError(
+            "Impossible to calculate scan points, please pass either the axes increment values or the number of scan points (n_images) per axis."
+            "For a 2D scan it is recommended that n_images is passed."
+        )
+
+    if len(axes_names) == 1 and rotation is True:
+        if not n_images:
+            n_images = round(abs(axes_starts[0] - axes_ends[0]) / axes_increments[0])
+
+        if axes_starts[0] != axes_ends[0] and axes_increments:
+            if axes_starts[0] > axes_ends[0]:
+                # Account for reverse rotation.
+                axes_ends[0] = axes_ends[0] + axes_increments[0]
+            else:
+                axes_ends[0] = axes_ends[0] - axes_increments[0]
+        elif axes_starts[0] != axes_ends[0] and not axes_increments:
+            inc = (axes_ends[0] - axes_starts[0]) / n_images
+            axes_ends[0] = axes_ends[0] - inc
+
+        spec = Line(axes_names[0], axes_starts[0], axes_ends[0], n_images)
+        scan_path = ScanPath(spec.calculate())
+
+    elif len(axes_names) == 1 and rotation is False:
+        if not n_images:
+            # FIXME This calculation still gives the wrong increment between scan points.
+            n_images = (
+                round(abs(axes_starts[0] - axes_ends[0]) / axes_increments[0]) + 1
+            )
+        elif type(n_images) is tuple and len(n_images) == 1:
+            # This is mostly a double paranoid check
+            n_images = n_images[0]
+
+        spec = Line(axes_names[0], axes_starts[0], axes_ends[0], n_images)
+        scan_path = ScanPath(spec.calculate())
+
+    else:
+        if not n_images:
+            # FIXME This calculation still gives the wrong increment between scan points.
+            n_images0 = (
+                round(abs(axes_starts[0] - axes_ends[0]) / axes_increments[0]) + 1
+            )
+            n_images1 = (
+                round(abs(axes_starts[1] - axes_ends[1]) / axes_increments[1]) + 1
+            )
+        elif len(n_images) == 1:
+            raise ValueError(
+                "Impossible to correctly calculate scan points from just the total number of images."
+                "Please either pass a tuple with the number of scan point per axis or the axes increments."
+            )
+        else:
+            n_images0 = n_images[0]
+            n_images1 = n_images[1]
+
+        if snaked is True:
+            spec = Line(axes_names[0], axes_starts[0], axes_ends[0], n_images0) * ~Line(
+                axes_names[1], axes_starts[1], axes_ends[1], n_images1
+            )
+        else:
+            spec = Line(axes_names[0], axes_starts[0], axes_ends[0], n_images0) * Line(
+                axes_names[1], axes_starts[1], axes_ends[1], n_images1
+            )
+        scan_path = ScanPath(spec.calculate())
+
+    return scan_path.consume().midpoints
 
 
 # Define the scan.
